Secure API Key
The script echoes the `user` and `api_key`, making it un-secure when
part of a deployment strategy (i.e. Capistrano recipe). I found the
most secured to be to store the `user` and the `api_key` on the
server where the application is being deployed as part of the setup
(i.e. /shared/path/rcfu/.RFUser, /shared/path/rcfu/.RFKey) and use
them as part of the deployment in a command like this:

	./scp-cloudfiles.sh.php \
		-u=`cat /shared/path/rcfu/.CFuser` \
		-k=`cat /shared/path/rcfu/.CFKey` \
		-c=<container> \
		-p=<path>